requirements.txt
Flask
flask-cors
tensorflow
numpy
pillow
spotipy
scikit-learn
matplotlib
gunicorn

